requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
